+++ frontend/src/app/admin/TenantDataAccess/TenantDataAccessForm.tsx
@@ -3,11 +3,10 @@ import React, { useEffect, useMemo, useState } from "react";
 import Button from "../../commonComponents/Button/Button";
 import RequiredMessage from "../../commonComponents/RequiredMessage";
 import TextInput from "../../commonComponents/TextInput";
-
 import OrganizationDropDown, {
   useOrganizationDropDownValidation,
   OrganizationOption,
-} from "./OrganizationDropDown";
+} from "../Components/OrganizationDropDown";
 
 const sortOrganizationOptions = (organizationOptions: OrganizationOption[]) =>
   Object.values(organizationOptions).sort((a, b) => {
